# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OmegaBot platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed, paused]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskList"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
  /llm/routes:
    get:
      operationId: listLlmRoutes
      tags: [llm]
      summary: List LLM routing rules
      responses:
        "200":
          description: Route list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmRouteList"
    post:
      operationId: createLlmRoute
      tags: [llm]
      summary: Create an LLM routing rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLlmRouteBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmRoute"
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, paused, awaiting_approval]
        priority:
          type: string
          enum: [low, medium, high, critical]
        adapter:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        runCount:
          type: integer
        lastRunAt:
          type: string
        tags:
          type: array
          items:
            type: string
        idempotencyKey:
          type: string
        writeSafe:
          type: boolean
      required: [id, name, status, priority, createdAt, updatedAt]
    TaskList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        total:
          type: integer
      required: [items, total]
    CreateTaskBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        adapter:
          type: string
        priority:
          type: string
        tags:
          type: array
          items:
            type: string
        idempotencyKey:
          type: string
        writeSafe:
          type: boolean
      required: [name]
    UpdateTaskBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
...
    LlmRoute:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        condition:
          type: string
        targetModelId:
          type: string
        targetModelName:
          type: string
        priority:
          type: integer
        enabled:
          type: boolean
        matchCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, condition, targetModelId, enabled, createdAt]
    LlmRouteList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/LlmRoute"
        total:
          type: integer
      required: [items, total]
    CreateLlmRouteBody:
      type: object
      properties:
        name:
          type: string
        condition:
          type: string
        targetModelId:
          type: string
        priority:
          type: integer
        enabled:
          type: boolean
      required: [name, condition, targetModelId]
...